+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unmerged] 
+- Allow zero amount transactions, removed flag
 - Progress bar enabled by default for only interactive mode
 
 ## [4.5.0] – 2022-06-10